refactor: extract common checkbox and radio-button core styles
