+++ packages/checkbox/src/vaadin-checkbox.js
@@ -9,8 +9,8 @@ import { ElementMixin } from '@vaadin/component-base/src/element-mixin.js';
 import { PolylitMixin } from '@vaadin/component-base/src/polylit-mixin.js';
 import { TooltipController } from '@vaadin/component-base/src/tooltip-controller.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import { checkboxStyles } from './styles/vaadin-checkbox-core-styles.js';
 import { CheckboxMixin } from './vaadin-checkbox-mixin.js';
-import { checkboxStyles } from './vaadin-checkbox-styles.js';
 
 /**
  * `<vaadin-checkbox>` is an input field representing a binary choice.
