+++ packages/field-base/src/styles/checkable-core-styles.js
@@ -3,9 +3,9 @@
  * Copyright (c) 2017 - 2025 Vaadin Ltd.
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
-import { css } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import { css, unsafeCSS } from 'lit';
 
-export const radioButtonStyles = css`
+export const checkable = (part, propName = part) => css`
   :host {
     display: inline-block;
   }
@@ -18,35 +18,37 @@ export const radioButtonStyles = css`
     -webkit-tap-highlight-color: transparent;
   }
 
-  .vaadin-radio-button-container {
+  .vaadin-${unsafeCSS(propName)}-container {
     display: grid;
     grid-template-columns: auto 1fr;
     align-items: baseline;
   }
 
-  [part='radio'],
+  [part='${unsafeCSS(part)}'],
   ::slotted(input),
+  [part='label'],
   ::slotted(label) {
     grid-row: 1;
   }
 
-  [part='radio'],
+  [part='${unsafeCSS(part)}'],
   ::slotted(input) {
     grid-column: 1;
   }
 
-  [part='radio'] {
-    width: var(--vaadin-radio-button-size, 1em);
-    height: var(--vaadin-radio-button-size, 1em);
+  /* Control container (checkbox, radio button) */
+  [part='${unsafeCSS(part)}'] {
+    width: var(--vaadin-${unsafeCSS(propName)}-size, 1em);
+    height: var(--vaadin-${unsafeCSS(propName)}-size, 1em);
     --_input-border-width: var(--vaadin-input-field-border-width, 0);
     --_input-border-color: var(--vaadin-input-field-border-color, transparent);
     box-shadow: inset 0 0 0 var(--_input-border-width, 0) var(--_input-border-color);
   }
 
-  [part='radio']::before {
+  [part='${unsafeCSS(part)}']::before {
     display: block;
     content: '\\202F';
-    line-height: var(--vaadin-radio-button-size, 1em);
+    line-height: var(--vaadin-${unsafeCSS(propName)}-size, 1em);
     contain: paint;
   }
 
@@ -58,20 +60,5 @@ export const radioButtonStyles = css`
     -webkit-appearance: none;
     width: initial;
     height: initial;
-  }
-
-  @media (forced-colors: active) {
-    [part='radio'] {
-      outline: 1px solid;
-      outline-offset: -1px;
-    }
-
-    :host([focused]) [part='radio'] {
-      outline-width: 2px;
-    }
-
-    :host([disabled]) [part='radio'] {
-      outline-color: GrayText;
-    }
   }
 `;
