+++ packages/radio-group/src/vaadin-radio-button.js
@@ -8,8 +8,8 @@ import { defineCustomElement } from '@vaadin/component-base/src/define.js';
 import { ElementMixin } from '@vaadin/component-base/src/element-mixin.js';
 import { PolylitMixin } from '@vaadin/component-base/src/polylit-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import { radioButtonStyles } from './styles/vaadin-radio-button-core-styles.js';
 import { RadioButtonMixin } from './vaadin-radio-button-mixin.js';
-import { radioButtonStyles } from './vaadin-radio-button-styles.js';
 
 /**
  * `<vaadin-radio-button>` is a web component representing a choice in a radio group.
